Refactor test scripts and update build configurations
- Added a new esbuild for:
  - bundling test runners (run tests on vscode + mocha runner)
  - bundling all tests in the project
- Some opinionated renamings / moving test related things around
- Changed output directory in tsconfig to avoid overriding the entire out directory - bundles tests also go there

Some poking around trying to deprecate jest as it is not used and mocha is more often used for vscode extension testing.
Though I assume jest was added to test GUI within extension.

This work was started because there are compilation issues with the current setup and neither of the tests are actually working so figured getting one to a working spot is a good start