Fix full-text search containg special characters (#876)
Previously the `ftsIndex.retrieve` would fail if any of the search terms contain characters like dots. This would be the case where you use the `folder` context provider searching using a folder containing dots.

Now we wrap each search term in double-quotes before joining them with `OR`.  This forces dots to be considered as part of the literal search term.

Co-authored-by: Lars Pagh <lp@airsupport.dk>