Fix markComplete removal (#874)
MarkComplete is called for every added, removed or changed item found. Therefore it seems odd (ignoring the performance perspective) that we try to remove all items at once. The other cases also only handle one single item at the time. Nevertheless the "remove" case did not behave as expected and did not remove anything from the tag_catalog.

With the code change we now handle one single item at the time and the item is removed from the tag_catalog as expected.

Co-authored-by: Lars Pagh <lp@airsupport.dk>