Add Language Name Param to AutocompleteTemplate (#1853)
Language name was previously added as a parameter only for string
autocomplete templates. This change provides this same information to
template functions that implement the AutocompleteTemplate interface.

Language name is explicitly supported for the CodeGeeX4 autocomplete
template, and is added here in this change. Explicit language awareness
could also be added for other templates in a future change, like the
holeFillerTemplate used for models without explicit completion support.