Move the bulk insert resolve into the COMMIT callback
db.db.serialize(...) guarantees that each statement runs in order. However, the function itself will terminate execution long before all the statements are executed. Therefore, we should only resolve the promise when the last COMMIT transaction has run. In the other cases, there will have been an error and we'll have call reject(err).

This also removed the need for the extra insert that was recently added so it was removed (in fact the extra insert was causing conflict errors).