Reorder stableCodeFimTemplate.stop order
Some providers have a limited number of STOPs that can be accepted
(e.g., NVIDIA NIMs, groq, etc). This means that the ordering of the
first 4 default STOPS is important. Starcoder has an interesting issue
where it emits <file_sep> as part of it's completion and having this
supplied as part of the 4 first stops helps to provide better
completions.

Signed-off-by: Adam Wallis <awallis@nvidia.com>