vscode: Ensure extension.js is rebuilt when deleted
Right now, the vscode-extension:continue-ui:build task
deletes extension.js as part of its prepackage.js script.

This wouldn't normally be a problem, because extension.js
is generated by the vscode-extension:esbuild, and
vscode-extension:continue-ui:build is a dependency of
vscode-extension:esbuild, implying it should always get run
first.

But, vscode-extension:esbuild is a background task, that stays
running after it's been started once. Future calls to run that
task become no-ops, because it's already active. This means running

vscode-extension:esbuild ->
  vscode-extension:continue-ui:build ->
    vscode-extension:esbuild

will lead to the last vscode-extension:esbuild not regenerating
extension.js.

In order to resolve that problem, this commit adds a new foreground
task into the build pipeline:

  vscode-extension:esbuild-notify

that tickles the watching background task to wake up and regenerate
extension.js