Improve Inline Tip
Remove background and outline of tip from SVG. This is required since the SVG content of a decoration cannot be taller than the line text height without causing the entire line to be pushed down.

Replace the box outline and background with CSS.

Fix spacing of text inside box. Add '+' symbols to shortcuts to make them clearer.