reapplyWithMyersDiff(): don't skip trailing blank lines
myersDiff inherits the jsdiff behavior for trailing \n, so would
consider "foo\n" => "foo" a changed single line, rather than a change
from "foo", "" to "foo" - this would result in a trailing blank line
in the highlighted region silently vanishing from the buffer.

Adding \n to the inputs so that lines are \n terminated rather than
separated fixes this.

To make myersDiff logically consistent, pass ignoreNewlineAtEof: true
to jsdiff. (This is actually not needed for this bug fix.)