Fixed unwantted selecting behavior for send-to-chat
The original behavior selects the entire first line even though it's
only partially selected in the editor. This fix is to get rid of the
extra selection from the first line and only send the exact part that
gets selected in the editor to the chat section. However, there is an
exception that if the context before the start of the selection
contains only whitespace, we want to include the wihitespace into the
selection because it's the indentation.
https://github.com/Granite-Code/granite-code/issues/43