Forbid file mentioning for large files
When the context items provided by user overflow the context length
and are truncated, it is very confusing for the user. This is a
first step that catches an common case at the least confusing
time - when the user is selecting a file. A more general check that
warns whenever the most recent message is truncated is something
that can be added later.
Granite-Code/granite-code#22