Standardize tooltip for chat section
Before this fix, tooltips were showing for some items and not for the others.
The inconsistency fells like a bug. In addidtion, the format of the tooltips
doesn't match the standard pattern in VS Code.

This fix provide tooltips for all the items in the Continue UI and formalize
the text format of each tooltip.
https://github.com/continuedev/continue/issues/4819