build: Remove libs.version.toml
I wanted to add a dependency (Sentry) and noticed that the plugins are described in a unusual way. I believe this is form over substance and overcomplicated, so I removed it. IMHO if we want to parameterize versions, it should be placed in `gradle.properties` or directly in `build.gradle.kts`.