test: Remove flaky test
Motivation: because `close` cancels the scope immediately without waiting for pending writes.

I initially wanted to fix the code to make `close` blocking until writes are flushed (with this PR), but I realized that we do not want to do that because the binary is meant to be terminated anyway and pending writes are not important. Implementation is correct. I decided to remove this flaky test.