Fix: Remove duplicate blocks and provide config warnings
This patch is to address the following cases that duplicate blocks appear in
the UI and confuse users:

  1. When users refer to a local block with "- uses", if the block is
     already in the default directories (e.g. ~/.continue/models or
     <workspacea>/.continue/models) and users don't override its name
     attribute, the duplicate block will appear in the UI.

  2. When users have duplicate block names among each block type in their
     assistant config.

  3. When the injected blocks have duplicate names with the assistant config
     for the corresponding block type.

The duplicate blocks in the UI is confusing, especially in the model
selection.

This patch add a blockDuplicationDetector to detect duplication. When a
duplication is detected, we remove the duplicate item and push a config
error to the result to warn the users that having duplicate names is not
the best practice to set up their config.

The detecting order: assistant config -> injected blocks

To detect duplication, all block types use "name" as the identifier except
for rules and context. For rules, if it's an rule object, we use the "name"
attribute. Otherwise, we use its string content as the identifier. For
context, the "provider" attribute is used as the identifier.

The detection only happens under assistant level, which means you can have
assistants with duplicate names.