fix: Apply proxy settings on the Continue process using environment variables (solves #5242) (#6752)
* fix: Apply proxy settings on the Continue process using environment variables (solves #5242)

This is a new and simplified implementation for passing proxy config.

It's based on environment variables instead of dedicated message in the protocol, however it has some downsides.

More details:
* This will not work if we use the binary with USE_TCP, because these env vars can only be set for the process.
* The implementation of the process handler has partially changed its behavior. From now on, manual restart (including Continue: Restart Process action) has no artificial delay/it's instant. This was a bug IMO that is fixed now.

To test this in a development environment, you need to remove USE_TCP from the run configuration.
Currently, we only track two proxy settings controls: whether it's enabled and the proxy address along with the port.
To test this, you need to go into the settings, change the state of those controls, then apply. The activity polls their values, and if it detects a change, the service restarts the binary, which reads the new setting on start.
There should always be only one process per IDE, regardless of how quickly the settings are changed or how often the binary is restarted manually.

* fix: misc.xml

* fix