fix: sanitize proxy settings + add NO_PROXY support + add unit tests (#7238)
Previously, proxy settings parsing was very "garbage-in, garbage-out". Since the HttpConfigurable class was written in Java, some settings could be null. We didn’t guard against that and simply concatenated proxy + port, which sometimes produced invalid values like "null:8080".

The implementation has been updated to sanitize + skip entries that cannot be used as env vars. In the worst case, the result is simply an empty map and the process env vars are unaffected.