fix: disable Next Edit cursor events when feature is disabled
Problem: To check if Next Edit is enabled, we need to make a request to the binary (+ read a settings JSON), which is suspendable/considered as "slow operation" and not suitable for blocking in events like caretPositionChanged/selection change.

Solution: This cache solves the problem with EDT violation exceptions when moving the cursor/selection.