fix: dispose JCEF JSQuery handler correctly
The JCEF browser wasn’t registered as a Disposable when held by the service. Service was disposed, but we had leak, because this caused JSQuery handlers to run on a disposed browser (browsers are disposed automatically by default according to docs), throwing exceptions when opening another project.

This fix registers both the browser and the handler properly. I tested it manually, and opening multiple projects no longer causes errors.

Note: I think it would be useful to have an e2e test at some stage that opens a project, closes it, and then opens another one. Such a test should catch this issue early.